requirements.txt
tensorflow==1.13.1
numpy>=1.13.3
absl-py